added extrapolation test
Noticed that if timePoint is after the last event, it's not so clear what to extrapolate and what not.
When extrapolating to the specified (maybe current) time, complete non-sense may result if the race
was long ago and not all leg fixes were yet received. If not extrapolating to the current point in time,
the question is to which time point to extrapolate. To the time point of the last event? To the time point
of the last fix in the (competitor's?) leg considered? And if such a more or less arbitrary selection
of a time point happens, ensure that this time point is then used throughout the calculations.