first rough GPS fix smoothening algorithm with a few tests in place
If one fix left and right has a computed speed of more than a threshold (currently
50 knots) that seems likely for the boat classes we're talking about, the fix is
removed from the smoothened view of the track.

The threshold obviously needs to become parameterizable. The subclass for
fixes that know their speed should refine this to take into account the speeds
as announced by the fixes themselves.