a half-way working version of an expand button in the table headers
Uses a CompositeCell to render a combination of a medal icon, race name and the expand/collapse button.
The challenge was to suppress column sorting when the expand/collapse button is clicked. I use
a somewhat shaky one-time suppress mechanism that tells the respective column to respond with "false"
when asked for sortability. After being asked the column becomes sortable again.