fixed serious bugs regarding storage of score corrections/max points reasons
One problem was the proper escaping of + characters which during URLDecoder.decode were
replaced by space. Another issue was that when directly calling setTrackedRace on a
RaceInLeaderboard column, the left-overs were not properly assigned. I fixed this using
a delegate pattern with a wrapper around the RaceInLeaderboard objects returned from
the leaderboard with delayed score correction / max points data. This, however, led to
a follow-up problem when constructing the keys used for the score correction / max points
maps: the wrappers were neither identical nor equal to the delegate. I now delegate the
key construction to the RaceInLeaderboard instance which gives the wrapper a chance to
use its delegate in the key constructed.