changed getOrCreateRaceDefinition such that TracTrac connector creates RaceDefinition and TrackedRace;
The problem was that AbstractReceiverWithQueue.getTrackedRace(Race race) waited for the RaceDefinition
and started fetching the TrackedRace immediately after it was unblocked by the notifyAll that
happened upon creating the RaceDefinition. At this point the RaceCourseReceiver may not have had
issued the creation of the TrackedRace. Hence, the lookup of the TrackedRace may have failed,
resulting in events being dropped instead of having gotten added to the TrackedRace.