fixed another deadlock issue;
The update of a GPSFixMovingTrackImpl was a synchronized method instead of, as the superclass does,
only synchronizing the update to the track itself but not the listener notifications.
The deadlock occurred because the GPS track notified the wind estimation track which in turn
iterates over the GPS track which requires synchronization.