trying a smarter way of dealing with the startToNextMarkDirection cache in TrackedRaceImpl;
the previous approach seems to have caused way too many task creations with a synchronized single-threaded
executor which had nothing to do but create, register and unregister cache invalidation listeners.
Now the class holds a map of cache invalidation listeners for the buoys of the first two waypoints.
This structure is updated when the waypoint sequence changes. Cache invalidations don't cause
changes in the listener structure.