Reproduces error in RaceTrackerTest.testStopTracking if executed after testInitialization
Run in "Run" mode. Order is important. When executed alone, it works. It seems that some caching
and timing problem is at fault. Added more logging, still not clear why the TrackedRegattaImpl
is re-used instead of created. I suppose the removal of the regatta doesn't remove the tracked regatta
properly.