clear server state entirely before starting to import initial load;
particularly registering the replica service for tracked race additions in order to auto-link them
to leaderboards causes trouble if the server state is not cleared before. Replicated tracked races may
then get auto-linked to old leaderboards and regattas in the server that happen to point to races with
equal RaceIdentifiers.

Also, it is important to become tolerant against a non-existing default leaderboard which doesn't exist
yet when tracked races are imported. The default leaderboard will be replicated together with other
leaderboards.