in LeadaerboardDTOCache distinguish cache objects by waitForLatestAnalyses parameter value to avoid inconsistencies; adjust time point to that of latest change
Have the cache adjust the time point of the request to that of the latest change to any of the attached races or the latest score correction to improve
cache hit ratio.

Eliminated RaceDTO.timePointOfLastEvent because it is just not interesting; timePointOfNewestEvent, however, is very interesting. Replaced a bunch
of usages of timePointOfLastEvent by the more adequate timePointOfNewestEvent.