no synchronized on TrackedRace anymore; it was useless and caused deadlocks during course updates in conjunction with mark passing updates;
Moved the synchronized writeObject(...) method to base class TrackedRaceImpl where the important data structures reside. Removed the
synchronized qualifier and rely only on the CouseImpl's lock. However, this is as unsafe as the synchronized was. What would actually
be required is a dedicated writeObject method that locks on each individual structure's lock when it writes it.