fixed simpleLeaderboardTest regarding the medal race consideration for the comparison with the discard thresholds
Due to a change in AbstractSimpleLeaderboardImpl.considerForDiscarding, medal races now (in line with the
comment for Leaderboard.considerForDiscarding) count for the comparison with the discarding thresholds even
though they are themselves never discarded. This has caused the flawed calculation of how many races to discard
in simpleLeaderboardTest to fail because it didn't consider the medal race during determining the number to
compare to the discarding thresholds.