Issues 1074 and 1075 resolved.
The Selenium runner provides now a description, which includes the used
browser and the capabilities (if defined). This also solves the problem,
that Eclipse was not able to show the correct results of a test run in
the JUnit view.

For a better support of the Page Object pattern, new annotations were
added, which are similar to the annotations already provided by
Selenium. However the new annotations are necessary, since the old
annotations work with enumerations which are mapped to classes and
which can not be extended. In contrast the new annotations expect a
class which allows the usage of new user-defined mechanisms
(implementations of the By interface) to locate elements within a
document. The only requirement is, that the implementing class must have
a public constructor which takes a String as parameter, whereby the
parameter describes the value to use for the search.