trying to fix bug 1312 by allowing LeaderboardDTOCache to use stale results of CrossTrackErrorCache and ManeuverCache;
The LeaderboardDTOCache is invalidated upon each GPS fix received; therefore, by and large, the LeaderboardDTOCache will
be invalidated each time the maneuver and cross track error caches will be invalidated. Only in very bizarre situations
such as the end of the race stopping all GPS fixes would the LeaderboardDTOCache keep some stale results for the very end
of the race. The likelihood seems sufficiently small to tolerate the risk.