use a single ThreadPool for all SmartFutureCache instances, 3*availableProcessors in size; shuffle multi-cache-invalidations;
when invalidating a SmartFutureCache for a number of keys, if this always happens in the same order, the first few keys
will be re-computed over and over again because cancelling them is not possible as they have already started, but a new
re-calculation needs to be triggered still. All later keys will be appended to the ordered queue but will keep getting
canceled because they haven't started their calculation yet. By shuffling the keys before triggering their invalidation,
all elements will fairly get re-calculated every now and then.