Tasks in the SmartFutureCache are now recycled with an extended update interval instead of canceled and re-scheduled if not yet started;
This solves bug 1314. A synchronized piece of code on the task attempts to update the interval in case the task hasn't read the interval
definition yet. If that doesn't succeed because the task has already consumed the old update interval definition, the existing future task
is left alone, and a new task with the extended update interval is created and scheduled.