really unlock read lock if counter == 1 or current thread is owner;
this takes care of the possibility that the thread to which locks are propagated gets the read lock itself and later
is propagated the read lock again. In this case, the receiving thread needs to first actually unlock its lock before
only decrementing the counter and relying on the propagating thread's lock ownership.