fixing bug 1349; DynamicTrackedRace held back triggers for the maneuver cache when the race was in state LOADING;
This was superceded by the change to suspend and resume the caches while the race is in state LOADING.
However, when suspended, triggers need to be sent to the cache, or else it won't start an update for
those keys when resumed.

Note that this commit does not fix the "side-show" issue of prematurely switching a race to TRACKING state
even if there are events left in its queue.