added locking for leaderboardCache map in LeaderboardDTOCache; we assume this will fix a non-working eviction;
The supposed problem could have been that in the presence of high load and a lot of concurrency, two removeEldestEntry calls
could have tried to remove the same entry. This will lead to one fewer entry being removed than put. Eventually, this leads to
a leak in the LeaderboardCacheDTO's internal map which should have been restricted to 10 entries. In the memory analyzer,
we saw 8296 entries in the 29er leaderboard.