Fixed issues with override mode in master data import
override mode made the whole import process overwrite events and regattas
again and again if they were (indirectly) contained in several leaderboard
groups. This is now fixed. The creation count object holds sets with
identifiers of the objects that have been created during the process.

also the number of created elements are validated in the test cases