Fixed a nasty bug where names of marks were not matching the ones configured in the TracTrac administration console. This bug hit the racecommittee hard because creation of a course has been almost impossible.
The solution is to make sure that single marks are created before the
corresponding gate. This avoids having strangely named marks created
upon gate initialization.

This commit also introduces hashCode for all mark and gate related
implementation. This is to ensure that only the Id is hashed because
this is the only really reliable base for object identity. It also
matches the semantics of getOrCreateMark(Serializable id).