refactored creation of ScoreCorrectionImpl and subclass instances so that it happens in leaderboard constructors;
Prepares for the linking of the score correction object back to its leaderboard in the score correction's constructor.
This will allow for the detection of race column ordering and hence the default application of score corrections
depending on time point if there are tracked races before or after that time point for a given competitor.