added a failing test case for the NPE during the deserialization of competitor-specific race log events;
The problem is that the competitor ID may not be resolvable in the receiving DomainFactory's competitor cache.
The root cause is that the receiver tries to resolve the race log event in the first place instead of simply
storing the race log event in the RaceLog. The solution will be not to trigger events when storing the race log
events into the RaceLog during master data import.