Refactored the AbstractFilterablePanel for the use in the data mining UI
The old AbstractFilterablePanel has been renamed to
LabeledAbstractFilterablePanel. Then I extracted everything except for the
label in a new super class, which is now called AbstractFilterablePanel.
Now the filterable panel can be used without a label before the filter
box. The LabeledAbstractFilterablePanel just calls the super constructor
and inserts the given label before the widget on index 0 (which is
currently the filter box).
I also changed the type of the display from CellTable to
AbstractCellTable, so that the filterable panel can be used with DataGrids
too (DataGrids are used for the selection tables in the data mining UI).