calculating maneuvers differently;
instead of looking at the leg type, only look at the current wind at the time and position of the maneuver and
calculate whether it's a tack and/or a jibe. If both, conclude that it was a penalty circle. If it's a mark passing,
now output two maneuvers: the MarkPassingManeuver *and* a bear away/head up/tack/jibe. This way we should now be able
to handle kiwi drops and jibe sets as a MarkPassingManeuver *and* a jibe/tack maneuver.

The good thing about this approach is that if it works we'll no longer depend on the leg type which may long have
changed in shifty wind conditions and only look at the boat and its course compared to the wind.