removed the tests that tested only StoreAndForward which is no longer needed for the new SwissTiming architecture;
SwissTiming will be providing an "SAP Gateway" which stores messages for us. We'll be able to query messages
even for a live race, using an "LSN" message sent to the gateway. Therefore, StoreAndForward as well as the
tests specific to this architecture are no longer required. The remaining tests are all green now.