Implemented a workflow to enrich the result with additional data
This was problematic, because a processor can have multiple result
receivers and every result receiver could add additional data. So I
implemented an additional result data builder, that handles the conflicts,
when a member of the additional data is set multiple times.
There's a remaining problem with the StringMessagesForTests, because it
isn't as rich as the string messages used in the productive environment.
When this is resolved, the query tests should be successful.