added a failing test case for the scenario described in bug 1854, comment #1;
Removing a race column does currently not remove its race logs in the database; neither does removing the entire
series or even regatta. The failing test case in TestRaceLogRemoval asserts (but so far fails in asserting) that
removing a race column and re-adding it will lead to a new, empty race log. The fix will naturally be to
add a method in MongoObjectFactory that removes a race log from the database which is called when a race column
is removed. This method should then probably also be invoked when an entire series or FlexibleLeaderboard is
removed.