fixing bug 896 by providing Timer.isInitialized() and not loading the leaderboard for an uninitialized timer;
the duplicate loading for the null time point in the Leaderboard.html entry point still persists; we would need
a smarter way of identifying that the offset adjustment which is causing the second load during the callback
wouldn't really required another load with null at basically the same time point, but the difficulty is that
then the timer's notion of live time and the leaderboard contents would deviate. Assuming that the null time
point is handled efficiently by the live leaderboard DTO cache, I think we can live with this.