to fix bug 1925, removed constraing in SailingServiceImpl.getAverageWindInfo to deliver wind fixes only if they are older than the time point of newest event;
See bug 1925; mainly, the problem seems to be that with an aggressive delay in live mode it can happen regularly that the client is posing
a request for wind data that is a bit newer than the newest event at that time for the current TrackedRace. In this case, although we would
extrapolate the tails, the wind display and advantage line would disappear, leading to a quite confusing flickering.