Cleaned the context hierarchy
Removed contexts for event and regatta, because an event or regatta
can't be used as atomar data type (they don't have enough data to make
serious queries possible). The highest context element is now the
HasTrackedRaceContext. Early filtering will still be possible, when the
query is working on RegattaLeaderboards, which will be the case for the
standard case.