added a first set of placeholders for a MVP architecture including an example, in the SSE environment;
The GWT compiler will strip any deliveries of the unused example. The code is not yet refactored. The
AppActivityMapper is still an example and should become something like a registry that we can use to
bind activities to places. The AppPlaceHistoryMapper is also example-specific and needs to be provided
by the respective application.