making testRaceTimeReplication more robust by waiting until at least the master has received the start of tracking time;
The problem here was that the abstract replication test set-up only waits until the tracked race has been received. Since the
use of the TracAPI the creation of the TrackedRace object doesn't require the successful receiving of the race timings. So we
need to wait for the arrival of the times separately.