addressing bug 2195 by at least remembering the lock that is locked for read;
The locksForKeys structure is written by the cache(K, V) method in SmartFutureCache. It was possible
(though very unlikely, which is probably why bug 2195 was observed only late and only once) that
the getOrCreateLockForKey(key) call in SmartFutureCache.get(...) produced the lock, then the cache(...)
method obtained the write lock just before the get(...) method was able to obtain the read lock. Then, when
the get() method obtained the read lock, the cache value might just have been removed, including the lock
that was removed. Then, when the reader correctly obtains null as the result and tries to unlock after read,
the lock obtained now is not the same anymore.