working on the failing testDynamicRaceLinking test case; the scoring scheme adjustments for non-sailed races broke this test;
Some competitors didn't have tracked ranks and previously were sorted to the end of the leaderboard. The new logic demands for
a low point scheme that they'd be sorted to the top of the leaderboard. While I think this is not useful (there should probably
be an exception for "no scores at all"), the test needs an adjustment of verdict for now.