fixing performance issue for maneuver queries in TrackedLegOfCompetitor;
The problem was that the calls required the latest results from the manuever cache.
With lots of load on the server, this can cause significant wait times for each
competitor, summing up for the hundreds of competitors that a big race can have.
This commit introduced a waitForLatest:boolean parameter to the corresponding
methods on TrackedLegOfCompetitor which is set to false by the currently existing
using client methods.