adjusted all versions of bundles in the workspace to 1.0.0-SNAPSHOT/.qualifier;
see https://wiki.wdf.sap.corp/wiki/display/LeanDI/Tycho+Versions+Plugin for an
in-depth explanation of what the tycho-version-plugin expects. The key issue is that
we wrap existing versioned components inside our workspace. This is not a supported
practice, at least not when working with the tycho-version-plugin. Therefore,
we have to live without versioning for those bundles unless we put them into
our own p2 repository in an OSGi-enabled form and reference them from our
target platform definition.