bug4110: introducing superclass of AbstractSimpleLeaderboardImpl for cache handling and DTO calculation;
The superclass doesn't, however, know about suppressing competitors; it
knows how to calculate a LeaderboardDTO and how to maintain a live and
non-live cache for those. It relies on specified but at this level
un-implemented functionality such as getCompetitorsFromBestToWorst to
compute the leaderboard contents. The plan now is to subclass this
with a leaderboard implementation delegating to an existing
RegattaLeaderboard but eliminating competitors not to be shown.

Change-Id: Iadffaa49e0c13e15bfe541c7f7a035d5bfa5a155