bug4687: Removed the consistency check when the query definition changes
This prevented notifying the listeners when the selected identity function changed. It isn't possible to check
if the retrieved data type is assignable to the statistics source type at the client side. This detects a wrong
inconsistency when an inherited function or a special function (like the identity function) is used as extraction
function. Removing the check shouldn't cause any errors, since this kind of consistency is ensured by the component
providers.