toJson() now needs JsonMode.RELAXED to remain compatible with org.simple.json.JSONParser;
otherwise, with the MongoDB 3.6 driver update, parsing documents
serialized using Document.toJson() will be unable to parse values of
type long properly. Those were usually represented either simply as the
number literal, or as NumberLong() constructor call. However, with the
new default behavior, they are represented as a nested Document with
a { $numberLong:...} format which cannot be cast to a Long object.
The RELAXED WriterSetting fixes this and reverts to the previous
version's default behavior.

Change-Id: I686439b7a332f2edddc6d0acb5423d80aecd1c0b