improved assertions in MPC; don't try to delete candidate from adjacent previous sequence;
the new invariant for the StationarySequenceBasedFilter states that no candidate can
be in more than one StationarySequence. Therefore, if it's found in one, no look
to the "left" adjacent sequence is required anymore.

Unfortunately, this still doesn't really explain what has happened at the SCL
as explained in bug 5026, comments 12 and following...