for MEDIA_TRACK creation, create ownership, then do check;
the special challenge here is that the media track's ID is generated
only by inserting it into the database. Only after that can the check be
performed whether the creation was allowed. The object then has to be
removed from the DB again, so it will have appeared shortly although the
user did not have the corresponding permission.

Change-Id: I4de6936f7cb947fc034b794c9ad581e764234b75