bug5098: including Waypoint UUID in the comparator disambiguates with Maps using Waypoint as key;
the problem with this bug was that Waypoint objects were used as keys
for two maps, but a Comparator was used for a TreeSet. When checking for
contains(...) on the TreeSet it seemed as if a Candidate with different
Waypoint object was already contained in the Set because the Waypoint ID
was not considered. However, for the candidate's waypoint no entries
existed in either map.

Change-Id: If3e71b4c4860d034feeb2c80497d4457f55343a0