return a comprehensive document from the ORC certificate import/assignment servlet;
the response now is a JSON document that shows how all requested
assignments worked out (OK / BOAT_NOT_FOUND / CERTIFICATE_NOT_FOUND). If
OK, the certificate ID is provided for the boat ID of the mapping.
Furthermore, all certificates parsed are output in the response. An
ORCCertificateJsonDeserializer should be able to read the objects in the
"certificates" array.

Change-Id: Ib3143d4908d82ae59658f6bdf7dc8cb229efbbc3