major refactoring of RaceLogResolver;
Introduced specialization that can also resolve a TrackedRace based on a
SimpleRaceLogIdentifier. This is then used to find TrackedRaces for race
log entries setting the implied wind to come from another TrackedRace.
With this, it doesn't matter whether the other tracked race has an ORC
Performance Curve ranking metric or not, or if the reference points to a
race slot that doesn't even have a tracked race attached. If a tracked
race is attached, it will be asked for its "reference implied wind."
Otherwise, the race log will be scanned, but if "own maximum implied
wind" is requested and there is no tracked race, null will be returned.

Change-Id: I14fe32f8a473281c9f489eb081544ae20c5b37f3