bug5194: started by fixing a bad bug causing outbound messaging to stop;
when de-registering a replica that caused one or more replicables to no
longer be observed, the outbound queue was prematurely closed and set to
null. For example, if one replica replicated the full server and another
one only the SecurityService, and then the full replica was
unregistered, sending security-related operations stopped even though a
replica was still registered.