bug5204: reverting regression effects of commit 5afaa1d4505d39372250ec1b11b8803e3271de6b;
trying to resize the ManeuverTable leads to a recursive resize effect
that eventually downsizes the maneuver table widget to height 0;
temporarily disabling the resize until a working solution has been
found. The effect will be that the resize handles may be pushed out of
view, e.g., when shrinking the containing window or when rotating a
small touch screen, hence changing aspect ratio.