windward distance to competitor farthest ahead now null before and constant after leg finish time point;
as reported by SwissTiming, when used through the RegattasResource's
/competitors/legs for time points before or after a competitor has
started/finished a leg, the gapToLeader-m value continued changing. It
turned out that the positions at timePoint were used for the
calculation, regardless of whether the competitor had finished the leg
already or hadn't even started the leg yet at that time point. With this
fix, for time points when the competitor hasn't started the leg yet,
null is returned; for time points after the competitor has finished the
leg, the time point when the competitor finished is used to determine
the leader and both, the leader's and the competitors position for the
distance calculation at that time point.