fixed NPE issues during loading an inconsistent local "replica" user store;
when replication produces a locally inconsistent version of the user
store, roles may not exist that users or groups reference. During
updates to the users somewhere in the landscape they get updated to each
server's local persistent user store. But not all roles may have made it
into that local store, and hence during the next attempt to load the
user store, loading groups referencing such a role, or loading users
that reference such a role failed with an NPE. The fix skips and logs
those unresolvable role definition references and assumes that a
subsequent initial load for the user store will fix things.