bug5322: fixed testing: exchange thread sync logic with ExecutionService and Future implementation.
The old implementation causes synchronization failures. It was possible to enter the synchronized block of the main thread before coming to the synchronized block of the child thread. Therefore the result was null (after the defined wait).
This behavior causes randomly assertion errors while building and testing the application.

Added org.objenesis dependency to manifest. This dependency was missing and therefore for local test execution was not possible (Igtimi tests).