bug5374: introduced asynchronous background metadata update;
a few tests fail now, probably because the accelerated fix insertion leaves less time
for actually committing the writes, and with write concern UNACKNOWLEDGED the client
may not necessarily read its own writes. Should we introduce sleep times? Or is there
a better way to have test cases wait for the completion also of unacknowledged writes?