bug4493: tackling the problem of tracker removal and re-creation cutting in between locks
see https://bugzilla.sapsailing.com/bugzilla/show_bug.cgi?id=4493#c5;
the solution first obtains the tracker under a read lock from
raceTrackersByRegatta, then releases the read lock and acquires the two
write locks in the correct order. Then the tracker is looked up under
the write locks, but it's removed if and only if it is object-identical
to the one found under the read lock because otherwise a new one may
have been created after the one to be removed was removed by a parallel
request.