bug5060: re-introduced SIMULATOR:READ for stand-along strategy simulator;
The permission is checked for the current server name as the object key.
Checks are performed client-side in the authorization wrapper which then
will display a "Sign in" page if the user is not logged in or lacks the
permission required. Furthermore, the SimulatorServiceImpl service
methods all check the permission on the server side.