bug5671bug5010: fixing issue with leaderboard detail premium columns;
They were erased from the map of columns available in the constructor of
ExpandableSortableColumn, and when the user upgraded to premium and
obtained the permissions necessary to see them, the map wasn't updated.
Now, the map continues to contain all columns available, regardless
their permission/action required, and only during expansion is the
permission check performed dynamically, when calculating the children
avaiable.