docker-11: activate the Java11 fragments for any Tycho target platform a always when Java 11 is detected;
The activation switch for the additional fragment dependency profile was
incorrect. It requested the with.not.android.relevant *system* property
to be set to "true", but that property is set only as a *Maven* property
and hence cannot be used for automatic profile activation. However,
activating always with Java-11 shouldn't be a problem when only the
fragments are added to the dependencies