improved logging in GenericCellTablePO.createEntry to find out more about test failures;
TestLeaderboardConfiguration.testDynamicRaceDeletion fails with 4%
probability when it cannot obtain the DataEntryPO objects for the races
in the series table after deleting one race column from that series. The
log so far ("Can't create DataEntry of type DataEntryPO") suggests that
something went wrong while trying to invoke the constructor.
I've added logging for all exceptions caught in the process of
stepping through the superclasses in case finally no constructor
was found or none terminated normally.