bug5127: special handling of double underscore in JSON. Double underscores in file name are encoded with hex representation and later decoded if read.
In some cases the JSON parser of Apples Safari on mobile devices cannot parse JSON value starting with a number and then double underscores. Usually there are no double __ in file name, but on Apple mobile devices the video filename is starting with digits following with two underscores, which will cause problems in JSON parser.