bug5838: when launching an instance, find security group by "Sailing Analytics App" Name tag;
additionally, set the subnet to which to deploy explicitly by
intersecting the subnets that match the desired availability zone with
the VPC ID of the security group used. This way, the VPC of the "Sailing
Analytics App"-name-tagged security group decides into which VPC the
instance(s) will be deployed.