jetty-9.4.53: trying to remove javax.xml.bind bundle (2.2.0) and jaxb-api bundle;
they seem to be of conflicting versions (2.2.0 vs. 2.3.0), and neither
should be required for Java8 builds where javax.xml.bind is part of the
JDK still. When this ever gets merged to master, merging into docker-17
should be expected to fail, and at least jaxb-api would then need to be
added again. We can only hope this will not cause problems with the
Tycho Surefire Selenium tests again, but at least starting from Java 11
the ambiguity with javax.xml.bind should be gone.