bug5774: The changes on AbstractGenericSerializableSettings are reverted and therefore another AbstractGenericSerializableSettingsWithContext was created instead. This prevent us to be forced to initialize every settings object with a security context even if it is not needed. Now we only use the ...WithContext in cases where we really need it to work with secured settings in the addChildSettings method.
There are still some places to work on e.g. where the secured DTO's are not in place when we init a settings with context. They are marked with TODO bug5774.
Also the secured ENUM setting needs some attention...

In LiveRaceWithRacemapAndLeaderBoardPresenterImpl there is already a change were we get a StrippedLeaderboardDTO to init context object correctly and which can be used for authorization check later.