don't invoke ./defineReverseProxyMappings.sh from sailing script;
it would require an initialized login shell with JAVA_HOME set properly,
but that's not the case when the /usr/local/bin/sailing script is
invoked through the sailing.service unit during the boot sequence. It is
no longer needed anyhow because the sailing instances no longer run an
Apache httpd server.