bug6059: moved Igtimi persistence test to new fragment; fixed OSGi websocket servlet setup;
the persistence test needs to be in the
com.sap.sailing.domain.igtimiadapter.persistence package; but with this
it has to be in a fragment that can legally extend the host bundle's
equally-named package. This was not the case with the
com.sap.sailing.domain.igtimiadapter.test fragment extending the host
bundle com.sap.sailing.domain.igtimiadapter. Moving to a separate
fragment extending the persistence bundle solved that problem.

The problem with WebSocket servlet tests was solved through the hint at
https://github.com/jetty/jetty.project/issues/7835.