don't send Igtimi AuthResponse without having seen the device serial number;
However, two firmware problems remain, and this doesn't solve them.
When sending a DGT in the successful AuthResponse, the device stops
sending the serial number in data messages. This would be fine because
now we delay the AuthResponse until having received the serial number at
least once. However, when the connection is interrupted thereafter, a
new AuthRequest on the newly established connection is attempted, but no
serial numbers are transmitted anymore. The firmware fails to realize
that with a new connection the server cannot know the serial number of
the connected device anymore. Therefore, we currently don't send the DGT
in the AuthResponse at all.

The other firmward problem is that there seems to be a problem, leading
to an error message:

  E -      : Cannot encode api message type. Only device group call supported

Riedel has so far failed to diagnose the root cause of this.