fix deadlock during RiotServerImpl initialization;
when the service is being initialized on a replica, queued operations
will be applied after receiving the initial load; at this time, the
SecurityServiceImpl may not yet be available, so asking for it can lead
to a deadlock. WebSocket connections, however, can not exist yet when
the RiotServerImpl hasn't announced it's ready yet. So accessing the
SecurityService only when there are active WebSocket connections should
solve this problem.